rename utils.dart -> options/angles.dart
