+++ lib/src/options/angles.dart
