+++ lib/steering.dart
@@ -4,7 +4,7 @@
 export 'src/steering/behavior.dart' show Behavior;
 export 'src/steering/kinematics.dart' show Kinematics;
 export 'src/steering/steerable.dart' show Steerable;
-export 'src/utils.dart' show angleToVector, vectorToAngle;
+export 'src/options/angles.dart' show angleToVector, vectorToAngle;
 
 // Kinematics implementations
 export 'src/steering/kinematics/basic_kinematics.dart' show BasicKinematics;
