+++ test/options/angles_test.dart
@@ -4,7 +4,7 @@ import 'package:radiance/steering.dart';
 import 'package:test/test.dart';
 import 'package:vector_math/vector_math_64.dart';
 
-import 'utils/close_to_vector.dart';
+import '../utils/close_to_vector.dart';
 
 void main() {
   group('Utils', () {
